+++ Sources/SentryObjCCompat/SentryObjCDataCollectionGraphQLCollectionOptions.swift
@@ -1,3 +1,4 @@
+#if SDK_V10
 // swiftlint:disable missing_docs
 #if SWIFT_PACKAGE
 internal import SentrySwift
@@ -39,3 +40,4 @@ public final class SentryObjCDataCollectionGraphQLCollectionOptions: NSObject {
 }
 
 // swiftlint:enable missing_docs
+#endif // SDK_V10